+++ src/app/AuthenticatedAppLayout.tsx
@@ -4,7 +4,9 @@ import { EncryptionProvider } from "../contexts/EncryptionContext";
 import { UserSettingsProvider } from "../contexts/UserSettingsContext";
 import { EncryptionGate } from "../components/EncryptionGate";
 import { DomainGate } from "../components/DomainGate";
+import { LocalCacheGate } from "../components/LocalCacheGate";
 import { DeviceActivityReporter } from "../components/DeviceActivityReporter";
+import { ErrorReporter } from "../components/ErrorReporter";
 import { UpdateProvider } from "../contexts/UpdateContext";
 import { AppShell } from "../components/layout/AppShell";
 import { ThemeProvider } from "../contexts/ThemeContext";
@@ -14,11 +16,13 @@ export function AuthenticatedAppLayout() {
     <>
       <AuthProvider>
       <DomainGate>
+      <LocalCacheGate>
       <UserSettingsProvider>
       <ThemeProvider>
       <EncryptionProvider>
         <EncryptionGate>
           <DeviceActivityReporter />
+          <ErrorReporter />
           <UpdateProvider>
             <AppProvider>
               <AppShell />
@@ -28,6 +32,7 @@ export function AuthenticatedAppLayout() {
       </EncryptionProvider>
       </ThemeProvider>
       </UserSettingsProvider>
+      </LocalCacheGate>
       </DomainGate>
     </AuthProvider>
     </>